STARTED A SWIMMING IN CIRCLES PROJECT
Did a lot of styling and structuring for this for fun project. I have hit a little burnout and facing some old demons. I always turn to Mac Miller for some guidance and decided to create a web page for voicing my opinions on his late pieces of work. RIP MALCOLM...